+++ tools/package-builder/src/core/cmd-opts.ts
@@ -43,6 +43,22 @@ export const BuildFlags = {
   }),
 };
 
+export const DockerFlags = {
+  'skip-docker-build': Flags.boolean({
+    env: envs.PL_DOCKER_SKIP_BUILD,
+    description: 'do not build docker images',
+    default: false,
+    required: false,
+  }),
+
+  'docker-autopush': Flags.boolean({
+    env: envs.PL_DOCKER_AUTOPUSH,
+    description: 'push docker images to registry after build. Enabled by default in CI builds.',
+    default: false,
+    required: false,
+  }),
+};
+
 export const DirHashFlag = {
   'full-dir-hash': Flags.boolean({
     env: envs.PL_PKG_FULL_HASH,