+++ etc/blocks/pool-explorer/block/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @milaboratories/milaboratories.pool-explorer
 
+## 1.0.277
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.191
+- @platforma-sdk/model@1.42.36
+
+## 1.0.276
+
+### Patch Changes
+
+- @milaboratories/milaboratories.pool-explorer.ui@1.1.190
+
 ## 1.0.275
 
 ### Patch Changes