+++ etc/blocks/pool-explorer/ui/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @milaboratories/milaboratories.pool-explorer.ui
 
+## 1.1.191
+
+### Patch Changes
+
+- Updated dependencies [b2e7c82]
+  - @milaboratories/helpers@1.7.0
+  - @platforma-sdk/ui-vue@1.42.44
+
+## 1.1.190
+
+### Patch Changes
+
+- Updated dependencies [b80a619]
+  - @platforma-sdk/ui-vue@1.42.42
+
 ## 1.1.189
 
 ### Patch Changes